requirements.txt
aiohttp==3.8.1
gunicorn==20.1.0
uvloop==0.16.0
Jinja2==3.1.2
asyncpg==0.25.0
aiohttp-cors==0.7.0
tortoise-orm==0.17.3

